blog-agent/pyproject.toml
[project]
name = "blog-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Happy-boy-001", email = "3298755711@qq.com" }
]
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.21.0",
    "chromadb>=1.5.8",
    "fastapi>=0.136.1",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.2.1",
    "langchain-text-splitters>=0.0.1",
    "langgraph>=1.1.9",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "loguru>=0.7.3",
    "pydantic>=2.13.3",
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "sqlalchemy>=2.0.49",
    "uvicorn>=0.46.0",
]

[project.scripts]
blog-agent = "blog_agent.main:main"

[build-system]
requires = ["uv_build>=0.11.4,<0.12.0"]
build-backend = "uv_build"
